Palentonga: tweak muts
Remove deformed body - there was no particular design goal there and,
with no bardings, it caused a weird power curve that was a little
stronger earlier and a little weaker later.

Add tail slap. This was always supposed to be in, I just completely
forgot it...

This commit doesn't include save compat, because this is an
experimental branch. Sorry, all.